feat: Add README.md file
README.md file was absent which resulted in lack of descriptive information about the repository. A README.md has been added to address this issue and to provide necessary details about the project. It will simplify the understanding for new contributors towards the project.

No breaking changes with this commit.

Issue: None